POLIO-1037: re-organise files
