+++ hat/assets/js/apps/Iaso/components/AlertModal/messages.ts
@@ -1,0 +1,8 @@
+import { defineMessages } from 'react-intl';
+
+export const MESSAGES = defineMessages({
+    close: {
+        defaultMessage: 'Close',
+        id: 'blsq.buttons.label.close',
+    },
+});
